cleaned repo, match reqs in issue
