+++ sniffer/Makefile
@@ -2,20 +2,17 @@ CC = gcc
 CFLAGS = 
 DEPS = 
 LDFLAGS = -lpcap
+OBJS = simple_sniff
 
 %.o: %.c $(DEPS)
 	$(CC) $(CFLAGS) -c -o $@ $<
 
-all: sniff simple_sniff
+all: $(OBJS)
 
 rebuild: clean all
 
 simple_sniff: simple_sniff.c
 	$(CC) -o $@ $^ $(LDFLAGS)
 
-sniff: simple_sniff2.c
-	$(CC) -o $@ $^ $(LDFLAGS)
-
-
 clean: 
-	@rm -f simple_sniff sniff
+	@rm -f $(OBJS)
